meson.build: create both shared and static libraries
shared lib will be used (in the future) when a standalone sdb is
created, so that it can be installed as a separate package/dependency
and used by other projects (not only radare2).

static lib can be used by other projects that include sdb as a
subproject (e.g. radare2, which includes the source code of sdb)